feat(lib::utils): move "parse_hex_color" to "config::theme" as that is the only place it is used
